global profile repositories moved to 'global.profile' package
