+++ src/main/java/egl/client/service/model/profile/GlobalCredentialsService.java
@@ -1,7 +1,7 @@
 package egl.client.service.model.profile;
 
 import egl.client.model.global.profile.GlobalCredentials;
-import egl.client.repository.profile.GlobalCredentialsRepository;
+import egl.client.repository.global.profile.GlobalCredentialsRepository;
 import egl.client.service.model.AbstractEntityService;
 import egl.core.model.profile.Profile;
 import org.springframework.stereotype.Service;
